some modify and bug fixed done

+++ admin/inc/delete_student.php
@@ -3,13 +3,17 @@
 
 
 
-        if( isset( $_GET['id'] )  ){
+        if( isset( $_GET['id'] ) &&  isset( $_GET['resultid'] )  ){
 
              $sid = $_GET['id'];
              $student_picture = $_GET['sphoto'];
+             $result_id = $_GET['resultid'];
 
              $sql = "DELETE FROM student_info where student_id='$sid' ";
-        $connection -> query($sql);
+             $connection -> query($sql);
+
+             $sql = "DELETE FROM student_results where student_id='$result_id' ";
+             $connection -> query($sql);
 
         unlink('../student_photos/'.$student_picture );
